feat: add keepalive util functions
