+++ internal/container/parser.go
@@ -5,150 +5,10 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/dagu-org/dagu/internal/digraph"
-	"github.com/dagu-org/dagu/internal/stringutil"
 	"github.com/docker/docker/api/types/container"
 	"github.com/docker/docker/api/types/mount"
-	"github.com/docker/docker/api/types/network"
 	"github.com/docker/go-connections/nat"
-	"github.com/go-viper/mapstructure/v2"
 )
-
-// NewFromContainerConfig parses digraph.Container into Container struct
-func NewFromContainerConfig(ct digraph.Container) (*Client, error) {
-	// Validate required fields
-	if ct.Image == "" {
-		return nil, ErrImageRequired
-	}
-
-	// Initialize Docker configuration structs
-	containerConfig := &container.Config{
-		Image:      ct.Image,
-		Env:        ct.Env,
-		User:       ct.User,
-		WorkingDir: ct.WorkDir,
-	}
-
-	hostConfig := &container.HostConfig{}
-	networkConfig := &network.NetworkingConfig{}
-	execOptions := &container.ExecOptions{}
-
-	// Parse volumes
-	if len(ct.Volumes) > 0 {
-		binds, mounts, err := parseVolumes(ct.Volumes)
-		if err != nil {
-			return nil, fmt.Errorf("failed to parse volumes: %w", err)
-		}
-		if len(binds) > 0 {
-			hostConfig.Binds = binds
-		}
-		if len(mounts) > 0 {
-			hostConfig.Mounts = mounts
-		}
-	}
-
-	// Parse ports
-	if len(ct.Ports) > 0 {
-		exposedPorts, portBindings, err := parsePorts(ct.Ports)
-		if err != nil {
-			return nil, fmt.Errorf("failed to parse ports: %w", err)
-		}
-		containerConfig.ExposedPorts = exposedPorts
-		hostConfig.PortBindings = portBindings
-	}
-
-	// Parse network mode
-	if ct.Network != "" {
-		networkMode := parseNetworkMode(ct.Network)
-		hostConfig.NetworkMode = networkMode
-
-		// If it's a custom network, add it to the endpoints config
-		if !isStandardNetworkMode(ct.Network) {
-			networkConfig.EndpointsConfig = map[string]*network.EndpointSettings{
-				ct.Network: {},
-			}
-		}
-	}
-
-	// autoRemove is the inverse of KeepContainer
-	autoRemove := !ct.KeepContainer
-
-	return &Client{
-		image:           ct.Image,
-		platform:        ct.Platform,
-		pull:            ct.PullPolicy,
-		autoRemove:      autoRemove,
-		containerConfig: containerConfig,
-		hostConfig:      hostConfig,
-		networkConfig:   networkConfig,
-		execOptions:     execOptions,
-	}, nil
-}
-
-// NewFromMapConfig parses executorConfig into Container struct
-func NewFromMapConfig(data map[string]any) (*Client, error) {
-	ret := struct {
-		Container     container.Config         `mapstructure:"container"`
-		Host          container.HostConfig     `mapstructure:"host"`
-		Network       network.NetworkingConfig `mapstructure:"network"`
-		Exec          container.ExecOptions    `mapstructure:"exec"`
-		AutoRemove    any                      `mapstructure:"autoRemove"`
-		Pull          any                      `mapstructure:"pull"`
-		Platform      string                   `mapstructure:"platform"`
-		ContainerName string                   `mapstructure:"containerName"`
-		Image         string                   `mapstructure:"image"`
-	}{}
-
-	md, err := mapstructure.NewDecoder(&mapstructure.DecoderConfig{
-		Result:           &ret,
-		WeaklyTypedInput: true,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed to create decoder: %w", err)
-	}
-	if err := md.Decode(data); err != nil {
-		return nil, fmt.Errorf("failed to decode config: %w", err)
-	}
-
-	var autoRemove bool
-	if ret.Host.AutoRemove {
-		ret.Host.AutoRemove = false // Prevent removal by sdk
-		autoRemove = true
-	}
-
-	pull := digraph.PullPolicyMissing
-	if ret.Pull != nil {
-		parsed, err := digraph.ParsePullPolicy(ret.Pull)
-		if err != nil {
-			return nil, err
-		}
-		pull = parsed
-	}
-
-	if ret.ContainerName == "" && ret.Image == "" {
-		return nil, ErrImageOrContainerShouldNotBeEmpty
-	}
-
-	if ret.AutoRemove != nil {
-		v, err := stringutil.ParseBool(ret.AutoRemove)
-		if err != nil {
-			return nil, fmt.Errorf("failed to evaluate autoRemove value: %w", err)
-		}
-		autoRemove = v
-	}
-
-	return &Client{
-		image:           ret.Image,
-		platform:        ret.Platform,
-		nameOrID:        ret.ContainerName,
-		pull:            pull,
-		containerConfig: &ret.Container,
-		hostConfig:      &ret.Host,
-		networkConfig:   &ret.Network,
-		execOptions:     &ret.Exec,
-		autoRemove:      autoRemove,
-	}, nil
-}
 
 // parseVolumes parses volume specifications into bind mounts and volume mounts
 func parseVolumes(volumes []string) ([]string, []mount.Mount, error) {
